% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Calculate_NCRMP_status_report_scores_GOM_Carib.R
\name{Calculate_NCRMP_status_report_scores_GOM_Carib}
\alias{Calculate_NCRMP_status_report_scores_GOM_Carib}
\title{Creates a dataframe of NCRMP status report benthic scores by region}
\usage{
Calculate_NCRMP_status_report_scores_GOM_Carib()
}
\arguments{
\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates a dataframe of NCRMP status report benthic scores by region
}
