% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compile_NCRMP_status_report_scores.R
\name{Compile_NCRMP_status_report_scores}
\alias{Compile_NCRMP_status_report_scores}
\title{Creates a dataframe of NCRMP status report benthic scores for all regions}
\usage{
Compile_NCRMP_status_report_scores(region)
}
\arguments{
\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates a dataframe of NCRMP status report benthic scores for all regions
}
