% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_Z_scores_cover.R
\name{NCRMP_calculate_Z_scores_cover}
\alias{NCRMP_calculate_Z_scores_cover}
\title{Creates percent cover dataframe}
\usage{
NCRMP_calculate_Z_scores_cover(
  region,
  indicator,
  min_year_ref,
  max_year_ref,
  min_year_current,
  max_year_current,
  reference_value,
  std,
  reef_type = "NULL"
)
}
\arguments{
\item{region}{A string indicating the region}

\item{indicator}{A string indicating benthic metric}

\item{min_year_ref}{A value indicating the first year of sampling to be included in the reference data}

\item{max_year_ref}{A value indicating the last year of sampling to be included in the reference data}

\item{min_year_current}{A value indicating the first year of sampling to be included in the current data}

\item{max_year_current}{A value indicating the last year of sampling to be included in the current data}

\item{reference_value}{A value indcating the reference mean}

\item{std}{A value indcating the reference std}

\item{reef_cat}{A string indicating reef classification}
}
\value{
A dataframe
}
\description{
Creates percent cover dataframe
}
