% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_Z_scores_mortality.R
\name{NCRMP_calculate_Z_scores_mortality}
\alias{NCRMP_calculate_Z_scores_mortality}
\title{Creates colony mortality summary dataframes}
\usage{
NCRMP_calculate_Z_scores_mortality(
  region,
  species_filter = "NULL",
  min_year_ref,
  max_year_ref,
  min_year_current,
  max_year_current,
  reference_value,
  std,
  reef_type
)
}
\arguments{
\item{region}{A string indicating the region}

\item{species_filter}{A string indicating whether to filter to a subset of species}

\item{min_year_ref}{A value indicating the first year of sampling to be included in the reference data}

\item{max_year_ref}{A value indicating the last year of sampling to be included in the reference data}

\item{min_year_current}{A value indicating the first year of sampling to be included in the current data}

\item{max_year_current}{A value indicating the last year of sampling to be included in the current data}

\item{reference_value}{A value indcating the reference mean}

\item{std}{A value indcating the reference std}

\item{reef_type}{A string indicating reef classification}
}
\value{
A dataframe
}
\description{
Creates colony mortality summary dataframes
}
