% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_demo_Zscores_FL.R
\name{NCRMP_make_weighted_demo_Zscores_FL}
\alias{NCRMP_make_weighted_demo_Zscores_FL}
\title{Creates weighted demo data}
\usage{
NCRMP_make_weighted_demo_Zscores_FL(
  input_ref,
  input_current,
  region,
  datatype,
  reef_type
)
}
\arguments{
\item{input_ref}{A dataframe}

\item{input_current}{A dataframe}

\item{region}{A string indicating the region}

\item{datatype}{A string indicating the datatype}

\item{reef_type}{A string indicating reef classification}
}
\value{
A dataframe
}
\description{
Creates weighted demo data
}
