% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_LPI_Zscores_FL.R
\name{NCRMP_make_weighted_LPI_Zscores_FL}
\alias{NCRMP_make_weighted_LPI_Zscores_FL}
\title{Creates weighted LPI data}
\usage{
NCRMP_make_weighted_LPI_Zscores_FL(
  input_ref,
  input_current,
  region,
  reef_type,
  indicator = "NULL"
)
}
\arguments{
\item{input_ref}{A dataframe}

\item{input_current}{A dataframe}

\item{region}{A string indicating the region}

\item{reef_type}{A string indicating reef classification}
}
\value{
A dataframe
}
\description{
Creates weighted LPI data
}
