% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_perform_ttest.R
\name{two_tail_t_test}
\alias{two_tail_t_test}
\title{Two Tail t-test}
\usage{
two_tail_t_test(dataframe, metric, alpha, return_dataframe = FALSE)
}
\arguments{
\item{dataframe}{A valid data.frame with, at a minimum, the following fields:
\itemize{
 \item{metric: }{specific metric calculated ("density", "old mortality", "coral cover", "macroalgae cover", "CCA cover")}
 \item{var: }{variance of calculated metric}
 \item{n: }{total number of PSUs sampled}
 \item{strat_num: }{number of strata sampled}
}}

\item{metric}{specific metric to be tested.  Must be in dataframe parameter ("density", "old mortality", "coral cover", "macroalgae cover", "CCA cover")}

\item{alpha}{alpha level to test metric (i.e 0.05, 0.01)}

\item{return_dataframe}{Default is FALSE
\itemize{
 \item{if FALSE: }{returns t-test significance to console}
 \item{if TRUE: }{returns input dataframe with additional fields (degrees of freedom(df), t_value,
 lower confidence interval(LCI), upper confidence interval(UCI) )}
}}
}
\value{
Either the t-test output to the console (if return_dataframe = F), or a dataframe containing values needed to conduct t-test
}
\description{
performs a 2-tailed t-test between two calculated metrics
}
\examples{
Run a t-test on dataset for density at alpha = 0.05, and return test output string
two_tail_t_test(dataset, "density", .05, return_dataframe = F)

Run a t-test on dataset for density at alpha = 0.01, and return the dataframe
two_tail_t_test(dataset, "density", .01, return_dataframe = T)

}
